Adjustments to ETL and read_file
Added a pre-configured schema for labels.csv, specifically to create a timestamp type instead of timestamp string. It didn't work, but it's there for experimentation. Also updated the column select in read_file to include object type. Finally I added some lines in BG2 that ingests the labels dataset and then converts the timestamp column from a string to a ts datatype.